Refactor user building
